+++ ess-20220222/src/Models/RemoveInstancesResponseBody/ignoredInstances.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Ess\V20220222\Models\RemoveInstancesResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class ignoredInstances extends Model
+{
+    /**
+     * @var string
+     */
+    public $code;
+
+    /**
+     * @var string
+     */
+    public $instanceId;
+
+    /**
+     * @var string
+     */
+    public $message;
+    protected $_name = [
+        'code' => 'Code',
+        'instanceId' => 'InstanceId',
+        'message' => 'Message',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->code) {
+            $res['Code'] = $this->code;
+        }
+
+        if (null !== $this->instanceId) {
+            $res['InstanceId'] = $this->instanceId;
+        }
+
+        if (null !== $this->message) {
+            $res['Message'] = $this->message;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Code'])) {
+            $model->code = $map['Code'];
+        }
+
+        if (isset($map['InstanceId'])) {
+            $model->instanceId = $map['InstanceId'];
+        }
+
+        if (isset($map['Message'])) {
+            $model->message = $map['Message'];
+        }
+
+        return $model;
+    }
+}